更新table；服务端非GET请求返回JSON增加id；删除RequestParser多余代码

+++ table/sys_Comment.sql
@@ -23,7 +23,7 @@ DROP TABLE IF EXISTS `Comment`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `Comment` (
-  `id` int(5) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `parentId` int(5) NOT NULL,
   `workId` int(5) NOT NULL,
   `userId` int(5) NOT NULL,
@@ -55,4 +55,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2017-01-01 19:21:37
+-- Dump completed on 2017-01-24 16:19:11
